Rename to useInvoiceable
