+++ components/fund-error.js
@@ -1,9 +1,9 @@
 import Link from 'next/link'
 import Button from 'react-bootstrap/Button'
-import { useAnonymous } from '../lib/anonymous'
+import { useInvoiceable } from './invoice'
 
 export default function FundError ({ onClose, amount, onPayment }) {
-  const anonPayment = useAnonymous(onPayment, { forceInvoice: true })
+  const createInvoice = useInvoiceable(onPayment, { forceInvoice: true })
   return (
     <>
       <p className='fw-bolder'>you need more sats</p>
@@ -12,8 +12,15 @@ export default function FundError ({ onClose, amount, onPayment }) {
           <Button variant='success' onClick={onClose}>fund wallet</Button>
         </Link>
         <span className='d-flex mx-3 font-weight-bold text-muted align-items-center'>or</span>
-        <Button variant='success' onClick={() => anonPayment(amount)}>pay invoice</Button>
+        <Button variant='success' onClick={() => createInvoice(amount)}>pay invoice</Button>
       </div>
     </>
   )
+}
+
+export const isInsufficientFundsError = (error) => {
+  if (Array.isArray(error)) {
+    return error.some(({ message }) => message.includes('insufficient funds'))
+  }
+  return error.toString().includes('insufficient funds')
 }
